[add] Mainコンポーネントの作成
