+++ src/features/layouts/index.ts
@@ -1,2 +1,3 @@
 export { Header } from './Header'
 export { Footer } from './Footer'
+export { Main } from './Main'
